+++ lib/mallorca_transit_services.dart
@@ -7,18 +7,23 @@
 library mallorca_transit_services;
 
 // APIs
-export 'src/api/departures.dart';
-export 'src/api/stations.dart';
-export 'src/api/route_line.dart';
-
-// Realtime
-export 'src/realtime/bus_position.dart';
-export 'src/realtime/bus_stopped.dart';
-export 'src/realtime/station_info.dart';
-export 'src/realtime/connection_close.dart';
-
-// Sockets
-export 'src/sockets/location_socket.dart';
+export 'src/api/departures_api.dart';
+export 'src/api/stations_api.dart';
+export 'src/api/route_lines_api.dart';
 
 // Messaging
 export 'src/messaging/transit_rss.dart';
+
+// Models
+export 'src/models/departure.dart';
+export 'src/models/station.dart';
+export 'src/models/route_line.dart';
+
+// Realtime models
+export 'src/models/realtime/bus_position.dart';
+export 'src/models/realtime/bus_stopped.dart';
+export 'src/models/realtime/station_info.dart';
+export 'src/models/realtime/connection_close.dart';
+
+// Sockets
+export 'src/sockets/location_socket.dart';