+++ lib/src/models/realtime/bus_position.dart
@@ -1,19 +1,8 @@
-/// Represents the position of a bus.
-/// It is received by the location socket and is used to track the bus.
-///
-/// The position includes the latitude, longitude, speed, and the timestamp of
-/// the position.
+/// Real-time bus position received from the location WebSocket.
 class BusPosition {
-  /// The latitude of the bus.
   final double lat;
-
-  /// The longitude of the bus.
   final double long;
-
-  /// The speed of the bus.
   final double speed;
-
-  /// The timestamp of the position.
   final DateTime timestamp;
 
   BusPosition(
@@ -27,7 +16,6 @@ class BusPosition {
     return 'BusPosition{lat: $lat, long: $long, speed: $speed, timestamp: $timestamp}';
   }
 
-  /// Converts a JSON map to a [BusPosition] object.
   factory BusPosition.fromJson(Map json) {
     return BusPosition(
         lat: json['lat'],